fix(): Build fix; LL replaced with HAL; Makefile created; Readme updated; Linked.ld created; STM32 libs imported; Project successfully compiled with output result in build_example.

+++ drivers/i2c_slave/i2c_slave.c
@@ -111,7 +111,7 @@ bool i2c_slave_init(I2C_HandleTypeDef *hi2c, uint8_t slave_addr)
     i2c_slave_handle = hi2c;
     
     /* Verify the handle is configured for slave mode */
-    if (i2c_slave_handle->Init.OwnAddress1 != (slave_addr << 1)) {
+    if (i2c_slave_handle->Init.OwnAddress1 != (uint32_t)(slave_addr << 1)) {
         return false;  /* Address mismatch */
     }
